Echo Generic Driver
Release B2

-New firmware for Gina3G & Layla3G that fixes pops and clicks with simultaneous MIDI
and audio I/O.  There is now only one DSP firmware image for both Gina3G and Layla3G.
-The new firmware does not support 8 kHz sampling
-Changed C3g and C3gDco to support the new firmware
-Added C3gDco::GetAudioMeters virtual method so that the correct meter values
        are returned for Gina3G and Layla3G
-Simplified CChannelMask; the mask is no longer stored byte-swapped on Macs and the
masks are at most 32 bits wide
-Moved the MakeDaffyDuck method from CEchoGals to a static member of CDaffyDuck;
the constructor for CDaffyDuck is now protected, so that you can only create new
ducks by calling MakeDaffyDuck.
-Various minor changes to avoid compiler warnings for AMD64 build
-OsSupportWdm has been rewritten to use the WDM DMA APIs


git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@14221 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/kernel/drivers/audio/echo/generic/CMonitorCtrl.cpp
@@ -135,12 +135,12 @@ ECHOSTATUS CMonitorCtrl::Init(CEchoGals *pEG)
 			{
 				m_Pans[wIndex] = 0;
 				m_PanDbs[wIndex].iLeft = 0;
-				m_PanDbs[wIndex].iRight = GENERIC_TO_DSP( ECHOGAIN_MUTED );
+				m_PanDbs[wIndex].iRight = (INT8) GENERIC_TO_DSP( ECHOGAIN_MUTED );
 			}
 			else
 			{
 				m_Pans[wIndex] = MAX_MIXER_PAN;
-				m_PanDbs[wIndex].iLeft = GENERIC_TO_DSP( ECHOGAIN_MUTED );
+				m_PanDbs[wIndex].iLeft = (INT8) GENERIC_TO_DSP( ECHOGAIN_MUTED );
 				m_PanDbs[wIndex].iRight = 0;
 			}
 			
@@ -249,7 +249,7 @@ ECHOSTATUS CMonitorCtrl::SetGain
 		else if (iGain < ECHOGAIN_MUTED)
 			iGain = ECHOGAIN_MUTED;
 	
-		m_Gains[ wIndex ] = GENERIC_TO_DSP( iGain );
+		m_Gains[ wIndex ] = (INT8) GENERIC_TO_DSP( iGain );
 		
 		//
 		// Gain has changed; store the notify
@@ -471,8 +471,8 @@ ECHOSTATUS CMonitorCtrl::SetPan(WORD wBusIn, WORD wBusOut, INT32 iPan)
 	//
 	//	Convert this pan setting into left and right dB values
 	// 		
-	m_PanDbs[wIndex].iLeft = GENERIC_TO_DSP( PanToDb(MAX_MIXER_PAN - iPan) );
-	m_PanDbs[wIndex].iRight = GENERIC_TO_DSP( PanToDb(iPan) );
+	m_PanDbs[wIndex].iLeft = (INT8) GENERIC_TO_DSP( PanToDb(MAX_MIXER_PAN - iPan) );
+	m_PanDbs[wIndex].iRight = (INT8) GENERIC_TO_DSP( PanToDb(iPan) );
 
 	//
 	// Store the notify
